Add factory manager class
The factory manager class will be used as the central control of all the
systems. It will receive the order data and process it. And then it can
do the planning and manage all the agvs.

Finished:
agv class & testing
sensors class & testing
factory manager class & testing

todo:
Embed sensor class in agv class
Control agvs with the factory manager